fastAPIDeploy/requirements.txt
onnx==1.17.0
onnxruntime==1.19.2
fastapi==0.104.0
uvicorn==0.32.0
Pillow==10.1.0
python-multipart==0.0.13